+++ backend-django/app/eric/core/views.py
@@ -15,12 +15,9 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.exceptions import MethodNotAllowed, ValidationError
 from rest_framework.response import Response
 
-import pkg_resources
 from drf_yasg.utils import swagger_auto_schema
 from git import Repo
 from memoize import memoize
-from pkg_resources._vendor.packaging.requirements import InvalidRequirement
-from pkg_resources._vendor.pyparsing import ParseException
 
 from eric.core.models import DisableSignals
 from eric.core.models.abstract import WorkbenchEntityMixin, get_all_workbench_models
@@ -32,28 +29,28 @@ from eric.versions.models.models import Version
 js_error_logger = logging.getLogger("js_errors")
 
 
-def get_pkg_metadata(pkgname, metadata_key="License"):
-    """
-    Given a package reference (as from requirements.txt),
-    return license listed in package metadata.
-    NOTE: This function does no error checking and is for
-    demonstration purposes only.
-    """
-    try:
-        pkgs = pkg_resources.require(pkgname)
-        pkg = pkgs[0]
-
-        for line in pkg.get_metadata_lines("METADATA"):
-            if ": " in line:
-                (k, v) = line.split(": ", 1)
-
-                if k.upper() == metadata_key.upper():
-                    return v
-
-    except FileNotFoundError:
-        return None
-
-    return None
+# def get_pkg_metadata(pkgname, metadata_key="License"):
+#     """
+#     Given a package reference (as from requirements.txt),
+#     return license listed in package metadata.
+#     NOTE: This function does no error checking and is for
+#     demonstration purposes only.
+#     """
+#     try:
+#         pkgs = pkg_resources.require(pkgname)
+#         pkg = pkgs[0]
+#
+#         for line in pkg.get_metadata_lines("METADATA"):
+#             if ": " in line:
+#                 (k, v) = line.split(": ", 1)
+#
+#                 if k.upper() == metadata_key.upper():
+#                     return v
+#
+#     except FileNotFoundError:
+#         return None
+#
+#     return None
 
 
 @memoize(timeout=60 * 60)
@@ -130,22 +127,22 @@ def oss_license_json(request):
     """
     license_json = []
 
-    from updatable import get_environment_requirements_list
-
-    installed_libraries = get_environment_requirements_list()
-
-    for library in installed_libraries:
-        one_license = dict()
-
-        one_license["key"] = library.split("==")[0]
-        try:
-            one_license["licenses"] = get_pkg_metadata(library, "License")
-            one_license["repository"] = get_pkg_metadata(library, "Home-page")
-
-            license_json.append(one_license)
-
-        except Exception:
-            pass
+    # from updatable import get_environment_requirements_list
+    #
+    # installed_libraries = get_environment_requirements_list()
+    #
+    # for library in installed_libraries:
+    #     one_license = dict()
+    #
+    #     one_license["key"] = library.split("==")[0]
+    #     try:
+    #         one_license["licenses"] = get_pkg_metadata(library, "License")
+    #         one_license["repository"] = get_pkg_metadata(library, "Home-page")
+    #
+    #         license_json.append(one_license)
+    #
+    #     except Exception:
+    #         pass
 
     return HttpResponse(json.dumps(license_json), content_type="application/json")